Refactor nifti module (#23)
