+++ nifti_test.go
@@ -2,21 +2,12 @@ package go2com
 
 import (
 	"fmt"
-	"github.com/okieraised/go2com/pkg/nifti/nifti2"
-	"github.com/okieraised/go2com/pkg/nifti/reader"
-	"github.com/stretchr/testify/assert"
 	_ "image/jpeg"
 	"testing"
-)
 
-func TestNii2(t *testing.T) {
-	assert := assert.New(t)
-	filePath := "/home/tripg/Documents/nifti/avg152T1_LR_nifti2.nii.gz"
-	niiReader, err := nifti2.NewNii2Reader(filePath)
-	assert.NoError(err)
-	err = niiReader.Parse()
-	assert.NoError(err)
-}
+	"github.com/okieraised/go2com/pkg/nifti/reader"
+	"github.com/stretchr/testify/assert"
+)
 
 func TestNii1(t *testing.T) {
 	assert := assert.New(t)
@@ -34,6 +25,7 @@ func TestNii1(t *testing.T) {
 	//filePath = "/home/tripg/Documents/nifti/ExBox11/structural_brain.nii.gz"
 	//filePath = "/home/tripg/Documents/nifti/JHU_MNI_SS_T1.nii.gz"
 	//filePath = "/home/tripg/Documents/nifti/avg152T1_LR_nifti2.nii.gz"
+	filePath = "/Users/TriPham/Downloads/avg152T1_RL_nifti.nii"
 
 	niiReader, err := reader.NewNiiReader(filePath)
 	assert.NoError(err)
