+++ pkg/nifti/reader/nifti1.h
